Домашннее задание 4.2. Управление схемами баз данных

+++ src/main/java/ru/hogwarts/school/entity/AverageAgeOfStudents.java
@@ -1,5 +1,5 @@
 package ru.hogwarts.school.entity;
 
 public interface AverageAgeOfStudents {
-    Integer getAverageAgeOfStudents();
+    Double getAverageAgeOfStudents();
 }
